Refactored more standard shared stuff into database module
