+++ bin/metabolomics_insert.py
@@ -50,36 +50,6 @@ class MetaBolInserter(DatabaseConnection):
     def get_person(self, person_id):
         return self._get(self.person_tbl, {'identifier': person_id })
 
-    def _get(self, table, info, return_primary_key=True):
-        s = table.select()
-        #print("TABLE: {} KEYS: {}".format(table, table.c.keys()))
-        for k, v in info.items():
-            #print("_get: {} {}".format( k, v) )
-            s = s.where( table.c[k] == v )
-        r = self.conn.execute(s).fetchone()
-
-        if not return_primary_key:
-            return None
-        if r:
-            return r.id
-        return None
-
-    def _insert(self, table, info, return_primary_key=True):
-        res = self.conn.execute(table.insert(), [ info ])
-        if not return_primary_key:
-            return
-        id = res.inserted_primary_key
-        return id[0]
-
-    def get_or_create(self, table, info, return_primary_key=True):
-        r = self._get(table, info, return_primary_key)
-        if r:
-            return r
-        try:
-            return self._insert(table,info,return_primary_key)
-        except sa.exc.SQLAlchemyError:
-            raise DatabaseError("Could not create {} from {}".format(table.name, json.dumps(info)))
-
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description="Insert metabolomics data")
     parser.add_argument('--file', type=str, required=True, help="TSV with metabolomics data")
