+++ bin/person_insert.py
@@ -19,35 +19,6 @@ class PersonInserter(DatabaseConnection):
             }
         )
 
-    def _get(self, table, info, return_primary_key=True):
-        s = table.select()
-        for k, v in info.items():
-            s = s.where( table.c[k] == v )
-        r = self.conn.execute(s).fetchone()
-
-        if not return_primary_key:
-            return None
-        if r:
-            return r.id
-        return None
-
-    def _insert(self, table, info, return_primary_key=True):
-        res = self.conn.execute(table.insert(), [ info ])
-        if not return_primary_key:
-            return
-        id = res.inserted_primary_key
-        return id[0]
-
-    def get_or_create(self, table, info, return_primary_key=True):
-        r = self._get(table, info, return_primary_key)
-        if r:
-            return r
-        try:
-            return self._insert(table,info,return_primary_key)
-        except sa.exc.SQLAlchemyError:
-            raise DatabaseError("Could not create {} from {}".format(table.name, json.dumps(info)))
-
-
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description="Insert a perons in the database")
     parser.add_argument('--identifier', type=str, required=True, help="Identifier for the person")
@@ -59,4 +30,3 @@ if __name__ == '__main__':
         PersonInserter().insert_person(args.identifier, args.group, args.sex)
     except DatabaseError as e:
         print("ERROR!: {}".format(e))
-
